+++ vscode/src/completions/inline-completion-item-provider.ts
@@ -43,6 +43,7 @@ import { isLocalCompletionsProvider } from './providers/experimental-ollama'
 import type { ProviderConfig } from './providers/provider'
 import { RequestManager, type RequestParams } from './request-manager'
 import { getRequestParamsFromLastCandidate } from './reuse-last-candidate'
+import { canReuseLastCandidateInDocumentContext } from './reuse-last-candidate'
 import { SmartThrottleService } from './smart-throttle'
 import {
     type AutocompleteInlineAcceptedCommandArgs,
@@ -147,6 +148,8 @@ export class InlineCompletionItemProvider
             isDotComUser: config.isDotComUser ?? false,
             noInlineAccept: config.noInlineAccept ?? false,
         }
+
+        autocompleteStageCounterLogger.setProviderModel(config.providerConfig.model)
 
         if (this.config.completeSuggestWidgetSelection) {
             // This must be set to true, or else the suggest widget showing will suppress inline
@@ -408,7 +411,21 @@ export class InlineCompletionItemProvider
                 const lastTriggeredPrefix = this.lastCandidate?.lastTriggerDocContext.currentLinePrefix
                 if (
                     this.lastCandidate &&
+                    // Remove the last candidate from cache only if it can be
+                    // used in the current document context.
+                    canReuseLastCandidateInDocumentContext({
+                        document,
+                        position,
+                        selectedCompletionInfo: context?.selectedCompletionInfo,
+                        lastCandidate: this.lastCandidate,
+                        docContext,
+                    }) &&
                     lastTriggeredPrefix !== undefined &&
+                    // TODO: consider changing this to the trigger point that users observe.
+                    // Currently, if the request is synthesized from the inflight requests with an earlier
+                    // trigger point, it is used here to decide if users wants to cancel the completion
+                    // but it's not obvious to the user where the trigger point is, which makes this
+                    // behavior opaque and hard to understand.
                     currentPrefix.length < lastTriggeredPrefix.length
                 ) {
                     this.handleUnwantedCompletionItem(