+++ vscode/src/services/StatusBar.ts
@@ -298,17 +298,21 @@ export function createStatusBar(): CodyStatusBar {
         // yellow status bar icon when extension first loads but login hasn't
         // initialized yet
         if (authStatus) {
+            if (authStatus.isOfflineMode) {
+                statusBarItem.text = '$(cody-logo-heavy) Offline'
+                statusBarItem.tooltip = 'Cody is in offline mode'
+                statusBarItem.backgroundColor = new vscode.ThemeColor('statusBarItem.warningBackground')
+                return
+            }
             if (authStatus.showNetworkError) {
                 statusBarItem.text = '$(cody-logo-heavy) Connection Issues'
                 statusBarItem.tooltip = 'Resolve network issues for Cody to work again'
-                // statusBarItem.color = new vscode.ThemeColor('statusBarItem.errorForeground')
                 statusBarItem.backgroundColor = new vscode.ThemeColor('statusBarItem.errorBackground')
                 return
             }
             if (!authStatus.isLoggedIn) {
                 statusBarItem.text = '$(cody-logo-heavy) Sign In'
                 statusBarItem.tooltip = 'Sign in to get started with Cody'
-                // statusBarItem.color = new vscode.ThemeColor('statusBarItem.warningForeground')
                 statusBarItem.backgroundColor = new vscode.ThemeColor('statusBarItem.warningBackground')
                 return
             }